Move model into database
