+++ app/src/main/java/com/gasperpintar/smokingtracker/database/dao/HistoryDao.kt
@@ -2,7 +2,7 @@ package com.gasperpintar.smokingtracker.database.dao
 
 import androidx.room.Dao
 import androidx.room.Query
-import com.gasperpintar.smokingtracker.model.CigarettesPerDay
+import com.gasperpintar.smokingtracker.database.model.CigarettesPerDay
 import com.gasperpintar.smokingtracker.database.entity.HistoryEntity
 import java.time.LocalDateTime
