+++ app/src/main/java/com/gasperpintar/smokingtracker/database/model/CigarettesPerDay.kt
@@ -1,4 +1,4 @@
-package com.gasperpintar.smokingtracker.model
+package com.gasperpintar.smokingtracker.database.model
 
 data class CigarettesPerDay(
    val dailySum: Int,
