+++ app/src/main/java/com/gasperpintar/smokingtracker/ui/fragment/achievements/AchievementEvaluator.kt
@@ -3,7 +3,7 @@ package com.gasperpintar.smokingtracker.ui.fragment.achievements
 import android.content.Context
 import com.gasperpintar.smokingtracker.R
 import com.gasperpintar.smokingtracker.database.entity.AchievementEntity
-import com.gasperpintar.smokingtracker.model.AchievementEntry
+import com.gasperpintar.smokingtracker.database.model.AchievementEntry
 import com.gasperpintar.smokingtracker.repository.AchievementRepository
 import com.gasperpintar.smokingtracker.repository.HistoryRepository
 import com.gasperpintar.smokingtracker.repository.NotificationsSettingsRepository
